More warning suppressions
	* lib-src/gnuslib.c (connect_to_unix_server): Use size_t for sz

	* src/ui/X11/event-Xt.c (x_reset_modifier_mapping): Add full
	prototype to auto functions

	* src/ui/glyphs-eimage.c (png_instantiate): use png_uint_32 in loop

	* src/ui/lwlib/xlwmenu.c (parameterize_string): add new offset to
	keep signedness in comparison

	* src/ui/X11/glyphs-x.c (x_subwindow_query_geometry): add prototyp

	* lib-src/gnuslib.c (connect_to_internet_server): Use size_t to
	compare with sizeof.

	* src/alloc.c (garbage_collect): ditto.
	* src/buffer.c (Fgenerate_new_buffer_name): ditto.
	* src/dumper.c (pdump_file_try): ditto.
	* src/dumper.c (pdump_load): ditto.
	* src/editfns.c (Fencode_time): ditto.
	* src/editfns.c (Fencode_btime): ditto.
	* src/editfns.c (Fcurrent_time_zone): ditto.
	* src/extents.c (print_extent_1): ditto.
	* src/mule/input-method-xlib.c (EmacsXtCvtStringToXIMStyles): ditto.
	* src/mule/input-method-xlib.c (describe_Window): ditto.
	* src/print.c (SXE_VSNPRINT_VA): ditto.
	* src/print.c (write_hex_ptr): ditto.
	* src/print.c (printing_major_badness): ditto.
	* src/process-unix.c (allocate_pty_the_old_fashioned_way): ditto.

	* src/process-unix.c (unix_open_network_stream): ditto.
	* src/process-unix.c (unix_open_network_server_stream): ditto.
	* src/process.c (make_process_internal): ditto.
	* src/strftime.c (add_num_time_t): ditto.
	* src/ui/TTY/objects-tty.c (Ffind_tty_color): ditto.
	* src/ui/TTY/objects-tty.c (Ftty_registered_color_list): ditto.
	* src/ui/TTY/redisplay-tty.c (term_get_fkeys_1): ditto.
	* src/ui/X11/device-x.c (read_locale_specific_resources): ditto.
	* src/ui/X11/device-x.c (x_init_device): ditto.
	* src/ui/X11/device-x.c (signal_if_x_error): ditto.
	* src/ui/X11/dialog-x.c (dbox_descriptor_to_widget_value): ditto.
	* src/ui/X11/event-Xt.c (x_keysym_to_emacs_keysym): ditto.
	* src/ui/X11/event-Xt.c (describe_event): ditto.
	* src/ui/X11/frame-x.c (color_to_string): ditto.
	* src/ui/X11/frame-x.c (x_set_initial_frame_size): ditto.
	* src/ui/X11/frame-x.c (Fx_window_id): ditto.
	* src/ui/X11/objects-x.c (truename_via_random_props): ditto.
	* src/ui/X11/scrollbar-x.c (x_create_scrollbar_instance): ditto.
	* src/ui/glyphs-eimage.c (tiff_error_func): ditto.
	* src/ui/glyphs-eimage.c (tiff_warning_func): ditto.
	* src/ui/lwlib/lwlib-Xaw.c (make_dialog): ditto.
	* src/ui/lwlib/lwlib-Xaw.c (make_dialog): ditto.
	* src/ui/lwlib/xlwgauge.c (GaugeExpose): ditto.
	* src/ui/lwlib/xlwgauge.c (MaxLabel): ditto.
	* src/ui/lwlib/xlwmenu.c (parameterize_string): ditto.
	* src/ui/lwlib/xlwtabs.c (XawTabsSetTop): ditto.
	* src/ui/redisplay.c (decode_mode_spec): ditto.
	* src/ui/redisplay.c (decode_mode_spec): ditto.
	* src/ui/redisplay.c (decode_mode_spec): ditto.

+++ src/buffer.c
@@ -705,7 +705,7 @@ even if a buffer with that name exists.
 	count = 1;
 	while (1) {
                 prt = snprintf(number, sizeof(number), "<%d>", ++count);
-	        assert(prt>=0 && prt < sizeof(number));
+	        assert(prt>=0 && (size_t)prt < sizeof(number));
 		gentemp = concat2(name, build_string(number));
 		if (!NILP(ignore)) {
 			tem = Fstring_equal(gentemp, ignore);
