+++ src/editfns.c
@@ -1352,7 +1352,7 @@ If you want them to stand for years in this century, you must do that yourself.
 			int sz = snprintf(tzbuf, sizeof(tzbuf), "XXX%s%d:%02d:%02d",
 					  "-" + (XINT(zone) < 0), abszone / (60 * 60),
 					  (abszone / 60) % 60, abszone % 60);
-			assert(sz >= 0 && sz < sizeof(tzbuf));
+			assert(sz >= 0 && (size_t)sz < sizeof(tzbuf));
 			tzstring = tzbuf;
 		} else {
 			error("Invalid time zone specification");
@@ -1427,7 +1427,7 @@ Like `encode-time' but return a big integer time instead.
 			int sz = snprintf(tzbuf, sizeof(tzbuf), "XXX%s%d:%02d:%02d",
 					  "-" + (XINT(zone) < 0), abszone / (60 * 60),
 					  (abszone / 60) % 60, abszone % 60);
-			assert(sz>=0 && sz < sizeof(tzbuf));
+			assert(sz>=0 && (size_t)sz < sizeof(tzbuf));
 			tzstring = tzbuf;
 		} else
 			error("Invalid time zone specification");
@@ -1566,7 +1566,7 @@ the data it can't find.
 			int sz = snprintf(buf, sizeof(buf), "%c%02d%02d", 
 					  (offset < 0 ? '-' : '+'),
 					  am / 60, am % 60);
-			assert(sz>=0 && sz < sizeof(buf));
+			assert(sz>=0 && (size_t)sz < sizeof(buf));
 			s = buf;
 		}
 		return list2(make_int(offset), build_string(s));
